セル開拓時のコストをdouble対応

+++ Assets/HK/AutoAnt/Scripts/Database/ConstantsCell.cs
@@ -32,7 +32,7 @@ namespace HK.AutoAnt.Database
         /// 開拓するためのコスト係数
         /// </summary>
         [SerializeField]
-        private int developCost = 10;
-        public int DevelopCost => this.developCost;
+        private double developCost = 10;
+        public double DevelopCost => this.developCost;
     }
 }
