+++ Assets/HK/AutoAnt/Scripts/GameControllers/Calculator.cs
@@ -12,7 +12,7 @@ namespace HK.AutoAnt.GameControllers
         /// <summary>
         /// セルを開拓するために必要なお金を返す
         /// </summary>
-        public static int DevelopCost(GameSystem gameSystem, Vector2Int position)
+        public static double DevelopCost(GameSystem gameSystem, Vector2Int position)
         {
             return gameSystem.Constants.Cell.DevelopCost * position.sqrMagnitude;
         }
